Create traveling-salesman
This program solves the Traveling Salesman problem. This program takes in the number of vertices and edges, then a list of edges with their costs, and finally the starting vertex. These costs are entered into their corresponding place in a 2-D vector array (i.e. an input of 0 2 13 sets costs[0][2] to 13). The program calls TSP, which runs through all permutations of the vertex set starting with the designated starting vertex. Within TSP, function cost is called which both tests if the path is valid and, if so, calculates the total cost of the path. cost returns the total cost to TSP. TSP then tests if the cost of the path just tested is smaller than the current minimum cost, and sets TSP_cost to the minCost and saves the path just tested. Once all permutations have been tested, the program outputs the path and cost of the cheapest trip. If no trip is possible, the program outputs no path and a cost of -1 signifying no path available.